modify mystl to cmakefile

+++ C++/STL/mystl/nested.h
@@ -4,7 +4,7 @@
 #include <bitset>
 #include <iostream>
 template <class T> class Nest { typedef typename T::iterator *ptr; };
-template <int N> void PrintBitset(std::bitset<N> const &bs) {
+template <long unsigned int N> void PrintBitset(std::bitset<N> const &bs) {
     std::cout << bs.template to_string<std::string::value_type,
                                        std::string::traits_type,
                                        std::string::allocator_type>()
